+++ blazeutils/datastructures.py
@@ -1,4 +1,5 @@
 from blazeutils.sentinels import NotGiven
+from blazeutils.containers import LazyDict
 
 # copied form webhelpers
 class BlankObject(object):
@@ -283,35 +284,6 @@ class LazyOrderedDict(OrderedDict):
         # attributes added after initialization are stored in _data
         else:
             self[item] = value
-
-class LazyDict(dict):
-    def __init__(self, **kwargs):
-        self._ld_initialized=kwargs.pop('_ld_initialize', True)
-        dict.__init__(self, **kwargs)
-
-    def __getattr__(self, attr):
-        if attr in self:
-            return self[attr]
-        raise AttributeError, "'%s' object has no attribute '%s'" \
-            % (self.__class__.__name__, attr)
-
-    def __setattr__(self, item, value):
-        # this test allows attributes to be set in the __init__ method
-        if self.__dict__.has_key('_ld_initialized') == False or self.__dict__['_ld_initialized'] == False:
-            self.__dict__[item] = value
-        # any normal attributes are handled normally when they already exist
-        # this would happen if they are given different values after initilization
-        elif self.__dict__.has_key(item):
-            self.__dict__[item] = value
-        # if there is a property, then set use it
-        elif self.__class__.__dict__.has_key(item) and isinstance(self.__class__.__dict__[item], property):
-            self.__class__.__dict__[item].__set__(self, value)
-        # attributes added after initialization are stored in _data
-        else:
-            self[item] = value
-
-    def __delattr__(self, name):
-        del self[name]
 
 from blazeutils.helpers import unique
 class UniqueList(list):