+++ changelog.rst
@@ -8,6 +8,8 @@ Changelog
 + added some utility functions for reST processing with docutils
 + added testing.FailLoader to help when testing failed imports
 + ensure_datetime() now takes a time_part argument
++ added HTMLAttributes to new containers module, moved LazyDict to containers but
+    left reference in datastructures.
 
 0.3.7 released 2011-12-15
 ==========================